qcacld: PMF: Fix to delete pmfSaQueryTimer during SubSystemRestart
The issue is, when SSR is triggered, we are not deleting pmfSaQueryTimer
because the function which actually deletes this timer is not getting
called during SSR.

This solution will fix this issue. As part of this solution, when ever
SSR is triggered, we will loop through each STA per BSSID and deletes
the pmfSaQueryTimer associated.

Change-Id: I01d6394d78b4ef93bdb5af24445c24d80d2fa8df
CRs-Fixed: 696054