qcacld: Avoid to re-submit RX URB in the suspend state
It exists a race condition that the URB is completed before usb_kill_urb is
called. Under this situation, the URB may be re-submitted although the HIF
is going to suspend. So when Linux is back from resume, it is possible that
URB is recieving data before HIF is totally resume. This has resulted in
data corruption sometimes.

Change-Id: I6d7c79398cd0985f15968f0ae1d0521f570e217d
CRs-Fixed: 753404