qcacld: hif: Implement an OOB interrupt handle mechanism.
Normally there are two options in the kernel to handle the SDIO
data, polling mode and inband interrupt mode.
In case inband interrupt mode isn't supported in the kernel which
will severely impact the throughput, the OOB interrupt handle
mechanism is implemented.
The basic idea of this OOB interrupt handle mechanism is to
register the OOB callback to the platform driver and then handle
the OOB interrupt in this OOB callback function.

Change-Id: If11ec60f1088c93112859938517163de8cb65473
CRs-Fixed: 937303
Git-commit: 1618afa173889a8cc3523bd0de5aa3e343ed909c