qcacld-2.0: Avoid integer overflow in wma_ndp_end_response_event_handler
The variable 'len_end_rsp' is a uint32 which can be overflowed if the
value of variable "event->num_ndp_end_rsp_per_ndi_list" is very large.
If the integer overflow occurs, finally it would lead to a heap overwrite
of heap object "end_rsp".

Check for len_end_rsp size , and ensure that an integer overflow does not
happen in wma_peer_info_event_handler.

Change-Id: I75cfaeacbee93fb27ad587feddfe298ca1d0d19b
CRs-Fixed: 2114348