qcacld-2.0: fix the ap mode ramdump and reinit issue when do LPM
this issue is caused by the wifi driver sent the
WMI_WOW_HOSTWAKEUP_FROM_SLEEP_CMDID cmd to fw after the WMI_WOW_ENABLE
return not success. In this case, our wifi worked as AP mode, and
there are one or more STAs connected our AP and there are data transfer
between AP and peer STAs, so the fw can not enter WOW and then the
WMI_WOW_ENABLE return not success.
For the fix, driver is disabled to send WMI_WOW_HOSTWAKEUP_FROM_SLEEP_CMDID
to fw if WMI_WOW_ENABLE return not success.

Change-Id: Ia550a9704056a82164e9b8bb359b1fef90a51167
CRs-Fixed: 2351670