qcacld-2.0: fix the wlan lpm issue on HQX3.0
This issue is caused by the NULL pointer access when do wlan driver
shutdown and do the process of wlan connecting operation in driver
at the same time. In this case, the driver is processing connecting
when do suspend, and our driver will do recovery when resume becasue
we have enabled the skip wow hostwakup(do not send wow wakeup to fw
and also do not need to wait 6 seconds) for HQX3.0. And there is a
timer started before suspend, and if it expired before this timer
destory when do PE stop, it will wakeup the MC thread and continue to
process it, but if some related resource freed durning the shutdown
before the MC thread stop, this NULL pointer access will happen.
So I skip to process messages in MC thread if we are doing shutdown
and this issue will not happen in this case.

Change-Id: Ifea29a43da8e7988a418b8d72ff44a39834d7522
CRs-fixed: 2418776