qcacld-2.0: fix agl suspend fail issue if wow enable fail on HQX3.0
this issue is caused by the skip wow hostwakeup on HQX3.0, this issue
will happen if wow enable fail when fw is busy on tx/rx running, and
then hif suspend fail but hdd(cfg80211) have already suspended, so wlan
driver shutdown/reinit will not happen but the hdd(cfg80211) resume will
not be called when skip wow hostwakeup enabled in this case.
So add an API to check whether driver reinit or not, if reinit not happen,
need to do hdd(cfg80211) resume.

Change-Id: I36a7aac30b9479bde7626b48efa4de28f3c5e021
CRs-fixed: 2419580