qcacld-2.0: Don't check for EXT_AUTH_AP_SUPPORT flag on new kernels
Propagation from qcacld-3.0 to qcacld-2.0

Backport flag CFG80211_EXTERNAL_AUTH_AP_SUPPORT is used to support
external auth in AP mode with older kernels. As the changes are
present in latest kernels, there is no need of this backport flag.

SAE PMK caching doesn't work on new kernels(e.g.: kernel-5.4) as
this backport flag is missing in latest kernels. So, avoid
checking for the same on new kernels.

Change-Id: Idc8858e7cff383e47196e2c94c66db40b5b98a3d
CRs-Fixed: 2807664